pyproject.toml
[project]
name = "support-assistant-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "yohannesalex",email = "yohannesalemayehu624@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "aiohttp (>=3.11.16,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"
alembic = "^1.15.2"

